Rename CSV interface with correct spelling
