+++ src/neuroconv/datainterfaces/text/csv/csvtimeintervalsinterface.py
